Add Levenshtein Distance
This is the iterative version of the levenshtein algo.

Tests
Tested equal, deletion, insertion and substitution. Each of the tests have been tested with an empty string and one change. Each of them have also been tested with multiple characters to cover edge cases. Null was not tested, or guarded as the method does not have nullable parameters.